Fixed dodge animation in midair, tuned camera movement speed

+++ src/h/drawingHandler.h
@@ -13,7 +13,7 @@ private:
     std::array<int, NLAYERS>& layerIndices;
     SDL_Renderer* renderer;
     Camera camera;
-    int shiftCount;
+    float shiftCount;
     int windowW;
     void removeInvalidComponents();
 public:
